Fix team ranks and ScoreIsTied() in Game VM
There was an off-by-one when checking team scores array. Red checked FFA
team, blue checked red team, and so on for green and yellow teams.

So when blue was winning the game thought score was tied (FFA and red
scores at 0) and did not check if blue reached the capture limit.