Fix losing race rank after finishing if following other player
If you finish the race and then follow another player (fire and then
rear fire to switch to SPECTATOR_FOLLOW) and then after other player
finishes and there is no one else to follow the game switches you
to SPECTATOR_FREE. This calls ClientBegin() which clears most of the
game client data, including the finishRaceTime. This puts the player
in last place on the scoreboard.