Fix out of range death events
Event EV_DEATH1 + sv_maxclients was sent when a player dies. This
is the wrong event. For sv_maxclients 10, CGame errors out because
EV_BREAK_GLASS isn't implemented.

EV_DEATH1 to 3 exist for different animations/sounds in Q3.
Specifically EV_DEATH1 + i is used. "i" for this was removed in
Q3Rally but "i" exists at a higher scope for iterating through
clients.