Fix bots going in reverse for no reason in races
On some maps bots went in reverse right away or got (almost to?) the
first checkpoint and then went in reverse forever. Race movement uses
derivatives on a curve but they seem to be invalid at fraction 0 and 1
so limit it to 0.05 to 0.95 (arbitrary).

This change fixes bots being stuck near the beginning of q3r_nightcity,
q3r_ridgeracer, and q3r_valley. The bots can complete the maps now.