Make mouse movement in menus match regular desktop movement
Use absolute mouse input instead of relative in the menus. Convert the
absolute position to relative movement in virtual "640x480 but aspect
correct" coords and preserve fractional movement. Synchronize the
position of the UI's menu cursor by forcing it to 0,0 and then sending
a delta of the current position in virtual coords.

The cursor can also freely move out of the window in the menu.