Fix camera not rotating smoothly
Reducing camera angle in damageYaw to 8-bit caused camera to be limited
to 256 angles and not rotate smoothly.

Restore setting it locally in higher precision as a separate variable
(damageAngles) but still keep it as 8-bit in damagePitch/damageYaw so
it networks correctly.

Issue introduced in "Fix sending angles in damagePitch/damageYaw".